fixup(account-management): address PR #1756 review findings
- Restore anti-starvation tiebreaker in `order_batch_leaf_first`
  (`(depth DESC, deletion_scheduled_at ASC, id ASC)`).
- Mark every `allow_all` / `scope_unchecked` site in `reads.rs` and
  `lifecycle.rs` with `// TODO(InTenantSubtree)` for greppability.
- Switch `redact_provider_detail` from `DefaultHasher` to `FnvHasher`
  so digests stay stable across Rust toolchain upgrades.
- Add a regression test that pins
  `DomainError::Internal::diagnostic` out of the public Problem
  envelope (currently enforced by `#[serde(skip)]` on `InternalV1`).
- Split `retention.rs` inline tests into `retention_tests.rs` to
  keep the inline block under DE1101's 100-line cap.
