+++ modules/system/account-management/account-management/src/infra/storage/repo_impl/reads.rs
@@ -145,6 +145,11 @@ pub(super) async fn is_descendant(
     let conn = repo.db.conn()?;
     let count = tenant_closure::Entity::find()
         .secure()
+        // TODO(InTenantSubtree): replace with the caller's narrowed
+        // scope once the predicate lands. Today every `allow_all` /
+        // `scope_unchecked` call site in this crate carries this
+        // marker so `rg "TODO(InTenantSubtree)"` greps the full
+        // bypass surface in one pass.
         .scope_with(&AccessScope::allow_all())
         .filter(
             Condition::all()
